replace model in route to fetch paginated filtered target profile summaries
