+++ api/lib/domain/usecases/find-paginated-filtered-target-profile-summaries-for-admin.js
@@ -1,0 +1,7 @@
+module.exports = function findPaginatedFilteredTargetProfileSummariesForAdmin({
+  filter,
+  page,
+  targetProfileSummaryForAdminRepository,
+}) {
+  return targetProfileSummaryForAdminRepository.findPaginatedFiltered({ filter, page });
+};
