--- api/lib/domain/usecases/find-paginated-filtered-target-profiles.js
@@ -1,3 +1,0 @@
-module.exports = function findPaginatedFilteredTargetProfiles({ filter, page, targetProfileRepository }) {
-  return targetProfileRepository.findPaginatedFiltered({ filter, page });
-};
