+++ api/lib/domain/usecases/index.js
@@ -139,6 +139,7 @@ const dependencies = {
   tagRepository: require('../../infrastructure/repositories/tag-repository'),
   TargetProfileForSpecifierRepository: require('../../infrastructure/repositories/campaign/target-profile-for-specifier-repository'),
   targetProfileRepository: require('../../infrastructure/repositories/target-profile-repository'),
+  targetProfileSummaryForAdminRepository: require('../../infrastructure/repositories/target-profile-summary-for-admin-repository'),
   targetProfileForUpdateRepository: require('../../infrastructure/repositories/target-profile-for-update-repository'),
   targetProfileShareRepository: require('../../infrastructure/repositories/target-profile-share-repository'),
   targetProfileWithLearningContentRepository: require('../../infrastructure/repositories/target-profile-with-learning-content-repository'),
@@ -266,7 +267,7 @@ module.exports = injectDependencies(
     findPaginatedFilteredOrganizationLearners: require('./find-paginated-filtered-organization-learners'),
     findPaginatedFilteredScoParticipants: require('./find-paginated-filtered-sco-participants'),
     findPaginatedFilteredSupParticipants: require('./find-paginated-filtered-sup-participants'),
-    findPaginatedFilteredTargetProfiles: require('./find-paginated-filtered-target-profiles'),
+    findPaginatedFilteredTargetProfileSummariesForAdmin: require('./find-paginated-filtered-target-profile-summaries-for-admin'),
     findPaginatedFilteredTargetProfileOrganizations: require('./find-paginated-filtered-target-profile-organizations'),
     findPaginatedFilteredUsers: require('./find-paginated-filtered-users'),
     findPaginatedParticipationsForCampaignManagement: require('./find-paginated-participations-for-campaign-management'),
